+++ internal/tool/process_tree_unix.go
@@ -1,0 +1,26 @@
+//go:build unix
+
+package tool
+
+import (
+	"context"
+	"os/exec"
+	"syscall"
+)
+
+func runCommandTree(ctx context.Context, command *exec.Cmd) error {
+	command.SysProcAttr = &syscall.SysProcAttr{Setpgid: true}
+	if err := command.Start(); err != nil {
+		return err
+	}
+	done := make(chan error, 1)
+	go func() { done <- command.Wait() }()
+	select {
+	case err := <-done:
+		return err
+	case <-ctx.Done():
+		_ = syscall.Kill(-command.Process.Pid, syscall.SIGKILL)
+		<-done
+		return ctx.Err()
+	}
+}